rpng: widen pass_geom arithmetic to size_t to prevent silent u32 wrap
rpng_pass_geom() computed `pitch` and `pass_size` in `unsigned int`
even though pass_size is stored through `size_t *`.  On inputs where
the intermediate products exceed 32 bits this wraps silently and
under-reports the size, which downstream causes an undersized
inflate buffer allocation and a heap buffer overflow during decode.

Concrete case: a 30000x30000 16-bit-per-channel RGBA PNG passes
the IHDR output-size cap (which gates on the RGBA-8 output buffer
of ~3.35 GiB, well under 4 GiB) but its 16bpc intermediate scanline
buffer needs 7.20 GiB.  With the old arithmetic this wrapped to
2.71 GiB (7,200,030,000 mod 2^32), the inflate buffer was allocated
at that wrapped size, and the inflate step would then write ~4.5 GiB
past the buffer end.  A crafted input of this shape has exploit
potential on any platform where the allocation happens to succeed.

Fix: lead each arithmetic chain with a `(size_t)` cast on the
first operand (typically ihdr->width or ihdr->depth) so the whole
expression promotes to size_t.  The local `bpp` and `pitch`
variables are widened to size_t to hold the result; the
out-parameters remain `unsigned *` and the narrowing cast at
assignment time is explicit.

On 64-bit builds this closes the wrap entirely — any computation
that would previously have wrapped now produces the correct value,
which downstream either allocates successfully (plenty of RAM) or
fails cleanly at malloc (returns NULL -> IMAGE_PROCESS_ERROR).

On 32-bit builds the pitch output can still be narrowed when the
real value exceeds UINT32_MAX, but that path is unreachable in
practice: reaching it requires an IHDR that the caller-side size
check also has to accept, and no currently-callable IHDR does.

Behaviour for all dimensions that previously computed correctly is
unchanged.  Verified empirically:

  100x100      RGBA-8  : pass_size unchanged (40100)
  23200x28000  RGBA-8  : pass_size unchanged (2,598,428,000)
  32767x32767  RGBA-8  : pass_size unchanged (4,294,737,923)
  30000x30000  RGBA-16 : old=2,905,062,704 (WRAPPED) new=7,200,030,000
  40000x20000  RGBA-16 : old=2,105,052,704 (WRAPPED) new=6,400,020,000

The two wrapping cases now report the correct size to the caller,
which will cleanly refuse the allocation on any real-world machine
instead of silently under-allocating.

Pre-existing latent bug, unrelated to recent changes; surfaced while
reviewing the IHDR-cap raise from 2 GiB to 4 GiB which widens the
range of inputs reaching this function.

+++ libretro-common/formats/png/rpng.c
@@ -957,41 +957,57 @@ static void rpng_pass_geom(const struct png_ihdr *ihdr,
       unsigned width, unsigned height,
       unsigned *bpp_out, unsigned *pitch_out, size_t *pass_size)
 {
-   unsigned bpp   = 0;
-   unsigned pitch = 0;
+   /* Perform pitch and pass_size arithmetic in size_t.  Previously these
+    * were done in unsigned int, which can silently wrap on 32-bit at a
+    * width of ~67M for 16bpc RGBA (pitch = width*8) or, more plausibly,
+    * at a pitch*height product exceeding ~4 GiB — reachable today with
+    * a 30000x30000 16bpc-RGBA image that passes the IHDR output-size
+    * cap (based on the RGBA-8 output buffer) but whose 16bpc intermediate
+    * scanline buffer is ~6.7 GiB.  A wrapped pass_size underallocates the
+    * inflate buffer and exposes a heap overflow during decode.
+    *
+    * The `(size_t)ihdr->width * ihdr->depth` leading term forces the
+    * whole expression to size_t width.  Callers with `unsigned *pitch_out`
+    * still receive a narrowed value — safe on 64-bit where size_t is 64
+    * bits, since realistic pitches fit comfortably.  On 32-bit targets
+    * pitch_out itself has no headroom beyond UINT32_MAX, but the caller
+    * won't reach any allocation using pitch if the IHDR check further
+    * down rejects such images for their overall size. */
+   size_t   bpp   = 0;
+   size_t   pitch = 0;
 
    switch (ihdr->color_type)
    {
       case PNG_IHDR_COLOR_GRAY:
-         bpp   = (ihdr->depth + 7) / 8;
-         pitch = (ihdr->width * ihdr->depth + 7) / 8;
+         bpp   = ((size_t)ihdr->depth + 7) / 8;
+         pitch = ((size_t)ihdr->width * ihdr->depth + 7) / 8;
          break;
       case PNG_IHDR_COLOR_RGB:
-         bpp   = (ihdr->depth * 3 + 7) / 8;
-         pitch = (ihdr->width * ihdr->depth * 3 + 7) / 8;
+         bpp   = ((size_t)ihdr->depth * 3 + 7) / 8;
+         pitch = ((size_t)ihdr->width * ihdr->depth * 3 + 7) / 8;
          break;
       case PNG_IHDR_COLOR_PLT:
-         bpp   = (ihdr->depth + 7) / 8;
-         pitch = (ihdr->width * ihdr->depth + 7) / 8;
+         bpp   = ((size_t)ihdr->depth + 7) / 8;
+         pitch = ((size_t)ihdr->width * ihdr->depth + 7) / 8;
          break;
       case PNG_IHDR_COLOR_GRAY_ALPHA:
-         bpp   = (ihdr->depth * 2 + 7) / 8;
-         pitch = (ihdr->width * ihdr->depth * 2 + 7) / 8;
+         bpp   = ((size_t)ihdr->depth * 2 + 7) / 8;
+         pitch = ((size_t)ihdr->width * ihdr->depth * 2 + 7) / 8;
          break;
       case PNG_IHDR_COLOR_RGBA:
-         bpp   = (ihdr->depth * 4 + 7) / 8;
-         pitch = (ihdr->width * ihdr->depth * 4 + 7) / 8;
+         bpp   = ((size_t)ihdr->depth * 4 + 7) / 8;
+         pitch = ((size_t)ihdr->width * ihdr->depth * 4 + 7) / 8;
          break;
       default:
          break;
    }
 
    if (pass_size)
-      *pass_size = (pitch + 1) * ihdr->height;
+      *pass_size = (pitch + 1) * (size_t)ihdr->height;
    if (bpp_out)
-      *bpp_out   = bpp;
+      *bpp_out   = (unsigned)bpp;
    if (pitch_out)
-      *pitch_out = pitch;
+      *pitch_out = (unsigned)pitch;
 }
 
 static void rpng_reverse_filter_adam7_deinterlace_pass(uint32_t *data,
